#41 Fix for SafeArray(String) constructor #111 m_pDispatch is not 0 if not attached Should really be mixing two fixes in the same commit but that's the way it goes

+++ jacob/jni/Dispatch.cpp
@@ -251,7 +251,7 @@ JNIEXPORT void JNICALL Java_com_jacob_com_Dispatch_release
   IDispatch *disp = (IDispatch *)num;
   if (disp) {
     disp->Release();
-    env->SetIntField(_this, jf, 0ll);
+    env->SetLongField(_this, jf, 0ll);
   }
 }
